backend/requirements.txt
flask
flask-cors
google-auth-oauthlib
google-auth
google-api-python-client
google-auth-httplib2
scikit-learn
mysql-connector-python
